chore: moved pages to organisms
